requirements.txt
tensorflow>=2.10
opencv-python
numpy
pandas
streamlit
matplotlib
scikit-learn
scikit-image
Pillow
tqdm
